// src/app/shared/models/user.ts
export class User {
    success?: string;
    data: {
        user:{
            category:string;
        }
    
    _id: string;
    name: string;
    password: string;
    mobileNumber: string;
    emailId: string;
    createdAt:string;
    __V:number;
    emailOtp:number;
    }
    token?: string;
}